% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/olson.R
\name{plot_tz}
\alias{plot_tz}
\title{Plot a timezone}
\usage{
plot_tz(tz)
}
\arguments{
\item{tz}{a valid time zone name. See \code{\link[=OlsonNames]{OlsonNames()}}}
}
\value{
a \code{ggplot2} object
}
\description{
Make a circular plot of a timezone, visualizing the UTC offset over the
course of the year, including Daylight Savings times
}
\examples{
plot_tz("America/Vancouver")
}
