% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tz-data.R
\docType{data}
\name{tz}
\alias{tz}
\title{Timezones}
\format{an `sf` object containing the timezones of the world}
\usage{
tz
}
\description{
These are from https://github.com/evansiroky/timezone-boundary-builder
and redistributed under the
[Open Data Commons Open Database License (ODbL)](http://opendatacommons.org/licenses/odbl/)
}
\keyword{datasets}
