% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/olson.R
\name{tz_list}
\alias{tz_list}
\title{Create a list of Time Zones}
\usage{
tz_list()
}
\value{
A data.frame of all time zones on your system. Columns:
\itemize{
\item tz_name: the name of the time zone
\item zone: time zone
\item is_dst: is the time zone in daylight savings time
\item utc_offset_h: offset from UTC (in hours)
}
}
\description{
Output a list of time zone names, with daylight savings time and utc offset
}
