% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tz_lookup.R
\name{tz_lookup_coords}
\alias{tz_lookup_coords}
\title{Lookup time zones of lat/long pairs}
\usage{
tz_lookup_coords(lat, lon)
}
\arguments{
\item{lat}{numeric vector of latitudes}

\item{lon}{numeric vector of longitudes the same length as \code{x}}
}
\value{
character vector the same length as x and y specifying the time zone of the points.
}
\description{
This can very quickly look up the time zone of many thousands of points,
however the underlying time zone map has been simplified to allow for this speed.
As such when a point is near a time zone boundary and not near a populated centre,
it may return the incorrect timezone. If accuracy is more important than speed,
try \code{\link[=tz_lookup_coords2]{tz_lookup_coords2()}}.
}
\examples{
tz_lookup_coords(42, -123)
tz_lookup_coords(lat = c(48.9, 38.5, 63.1, -25), lon = c(-123.5, -110.2, -95.0, 130))
}
