% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tz_lookup.R
\name{tz_lookup_sf}
\alias{tz_lookup_sf}
\title{Lookup timezones of points}
\usage{
tz_lookup_sf(x, crs = NULL)
}
\arguments{
\item{x}{either an \code{sfc} or \code{sf} points object}

\item{crs}{the coordinate reference system: integer with the EPSG code, or character with proj4string.
If not specified (i.e., \code{NULL}) and \code{x} has no existing \code{crs}, EPSG: 4326 is assumed (lat/long).}
}
\value{
character vector the same length as \code{x} specifying the Timezone of the points.
}
\description{
Lookup timezones of points
}
\examples{
if (requireNamespace("sp")) {
data(meuse, package = "sp")
meuse_sf = sf::st_as_sf(meuse, coords = c("x", "y"), crs = 28992, agr = "constant")
tz_lookup_sf(meuse_sf)
}

}
