% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/olson.R
\name{tz_plot}
\alias{tz_plot}
\title{Plot a time zone}
\usage{
tz_plot(tz)
}
\arguments{
\item{tz}{a valid time zone name. See \code{\link[=OlsonNames]{OlsonNames()}}}
}
\value{
a \code{ggplot2} object
}
\description{
Make a circular plot of a time zone, visualizing the UTC offset over the
course of the year, including Daylight Savings times
}
\examples{
tz_plot("America/Vancouver")
}
