% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tz_lookup.R
\name{tz_lookup}
\alias{tz_lookup}
\title{Lookup timezones of points}
\usage{
tz_lookup(x, y, crs = NULL)
}
\arguments{
\item{x}{either an `sfc` or `sf` points object, or a numeric vector of x coordinates}

\item{y}{if `x` is a numeric vector of x coordinates, `y` must be a numeric vector of
y coordinates the same length as `x`. Ignored if `x` is an `sf` or `sfc` object.}

\item{crs}{the coordinate reference system: integer with the EPSG code, or character with proj4string.
If not specified (i.e., `NULL`) and `x` and `y` are numeric vectors, or `x` is an `sf` or `sfc` object
with no existing `crs`, EPSG: 4326 is assumed (lat/long).}
}
\value{
character vector the same length as specifying the Timezone of the points.
}
\description{
Lookup timezones of points
}
\examples{
if (requireNamespace("sp")) {
data(meuse, package = "sp")

# numeric:
tz_lookup(meuse$x, meuse$y, 28992)

# With a sf object:
meuse_sf = sf::st_as_sf(meuse, coords = c("x", "y"), crs = 28992, agr = "constant")
tz_lookup(meuse_sf)
}

}
