% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tz_lookup.R
\name{tz_lookup}
\alias{tz_lookup}
\title{Lookup timezones of points}
\usage{
tz_lookup(lat, lon)
}
\arguments{
\item{lat}{numeric vector of latitudes}

\item{lon}{numeric vector of longitudes the same length as \code{x}}
}
\value{
character vector the same length as x and y specifying the Timezone of the points.
}
\description{
Lookup timezones of points
}
\examples{
tz_lookup(42, -123)
tz_lookup(lat = c(48.9, 38.5, 63.1, -25), lon = c(-123.5, -110.2, -95.0, 130))
}
