% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tz_lookup.R
\name{tz_lookup2}
\alias{tz_lookup2}
\title{A more accurate (but slower) version of tz_lookup}
\usage{
tz_lookup2(x, crs)
}
\arguments{
\item{x}{either an \code{sfc} or \code{sf} points or \code{SpatialPoints(DataFrame)} object}

\item{crs}{the coordinate reference system: integer with the EPSG code, or character with proj4string.
If not specified (i.e., \code{NULL}) and \code{x} has no existing \code{crs}, EPSG: 4326 is assumed (lat/long).}
}
\value{
character vector the same length as \code{x} specifying the time zone of the points.
}
\description{
A more accurate (but slower) version of tz_lookup
}
